Switch CompletionRecord to a namedtuple
I was worried that creating CompletionRecord objects all over the place
would be slow and greedy for memory (and require too frequent garbage
collection).

The 'namedtuple' collection does everything we need, for a much lighter
cost. (We still are probably garbage collecting too much, but this is a
lot more simple an object.)