Fix deferred grammar functions in the parse
The functions used in the parse grammar often defer to their child
nonterminals to do the work. This originally was just the first child,
but it really should have been "sole nonterminal".

This change also allowed a check for that "sole" bit, which means a few
missing functions from the
	StatementList: StatementList StatementListItem
production got noticed, and, as a result, added.