Alter the core functions to take any args
The first two functions built in the intrinsic stuffer (the one that
does nothing, and the one that throws a TypeError) are altered to take
any combination of arguments. (They don't depend on args anyway, and
this deals with my forgetting about new_target, and in addition means
that they satisfy any calling signature, which is probably a good
thing.)

Also fixed a typo in a function name.