Add a cache to the token scanner
Turns out we call token() for the same source text a **lot**. We were
spending all of our time doing regex matches that were the same as the
last time we did them.

This change adds a simple cache in there so that if we hit the tokenizer
for the same source position and same goal, we just return the
previously computed token.