Change init of _schars in regex to be speedier
My perf tests were showing that I was spending a huge amount of time
scanning the unicode code range to find Zs chars every time we launched.
But it doesn't ever change unless the unicode database changes, so it
feels a lot more appropriate for build time.

Now _schars is a pregenerated string. Old code is there (but commented
out), in case we ever need to regenerate it.