Working harder to get python to match dtoa.
I suspect that python doesn't actually use dtoa in its internals. In any
event, I can't figure out how many digits are ever actually significant
in it's double-to-string conversion (I don't think it's always the
same). So we now do string conversion for what we get back from the
formatter, as well as +/- 1 from that value. If we get the same float
back we came in with, we use the form with the fewer number of
characters.

This _seems_ to match what ECMAScript does.