Alter the function checks so snoop works
The @snoop decorator gives a function a different argspec, so the checks
in BindBuiltinFunction would always fail. This change tries to detect
that and works down into the wrapped function to get the _real_ argspec.
This can still be fooled if you work at it, but the idea is to let me
@snoop a troublesome function without having to disable the checks all
the time.